requirements.txt
Django==1.11.20
djangorestframework==3.9.2
Pillow==5.4.1
psycopg2-binary==2.7.7
